rename : DBconfig 으로 폴더 이동
