+++ src/main/java/com/fourclover/clobee/config/dbConfig/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.fourclover.clobee.config;
+package com.fourclover.clobee.config.dbConfig;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
